Make a separate package for conan.cmake
This used to be part of cmake-modules, but is really just tracking
an upstream package.

Note that it is not useful to build_requires this in a conanfile.txt;
by the time if you're installing a conanfile, you don't need it anymore.

But it's still convenient to distribute and version the files through
artifactory and use `conan install` and the ConanFile.deploy() method
to register it with CMake find_package